+++ src/app/smarthome/lib/services/api/base.dart
@@ -5,9 +5,8 @@ import 'dart:io';
 
 import 'package:Smarthome/constants/api.dart';
 import 'package:Smarthome/constants/constants.dart';
-import 'package:Smarthome/core/certificates.dart';
+import 'package:Smarthome/core/certificate_validation.dart';
 import 'package:Smarthome/models/rest_resource.dart';
-import 'package:Smarthome/redux/actions.dart';
 import 'package:Smarthome/redux/store.dart';
 
 class HTTP {
